+++ docs/src/modules/components/AppFrame.js
@@ -35,6 +35,7 @@ import { pathnameToLanguage } from 'docs/src/modules/utils/helpers';
 import RtlContext from 'docs/src/modules/utils/RtlContext';
 import { useChangeTheme } from 'docs/src/modules/components/ThemeContext';
 import PageContext from 'docs/src/modules/components/PageContext';
+import { useTranslate } from 'docs/src/modules/utils/i18n';
 
 const LOCALES = { zh: 'zh-CN', pt: 'pt-BR', es: 'es-ES' };
 const CROWDIN_ROOT_URL = 'https://translate.material-ui.com/project/material-ui-docs/';
@@ -146,7 +147,7 @@ const styles = (theme) => ({
 function AppFrame(props) {
   const { children, classes, disableDrawer = false } = props;
   const theme = useTheme();
-  const t = useSelector((state) => state.options.t);
+  const t = useTranslate();
   const userLanguage = useSelector((state) => state.options.userLanguage);
   const { rtl, setRtl } = React.useContext(RtlContext);
