+++ docs/src/modules/components/Head.js
@@ -3,9 +3,10 @@ import NextHead from 'next/head';
 import { useRouter } from 'next/router';
 import PropTypes from 'prop-types';
 import { useSelector } from 'react-redux';
+import { useTranslate } from 'docs/src/modules/utils/i18n';
 
 export default function Head(props) {
-  const t = useSelector((state) => state.options.t);
+  const t = useTranslate();
   const {
     card = 'https://material-ui.com/static/logo.png',
     children,
