+++ docs/src/modules/components/Notifications.js
@@ -17,6 +17,7 @@ import ListItem from '@material-ui/core/ListItem';
 import Divider from '@material-ui/core/Divider';
 import { getCookie } from 'docs/src/modules/utils/helpers';
 import { ACTION_TYPES } from 'docs/src/modules/constants';
+import { useTranslate } from 'docs/src/modules/utils/i18n';
 
 const useStyles = makeStyles((theme) => ({
   paper: {
@@ -46,7 +47,7 @@ export default function Notifications() {
   const [open, setOpen] = React.useState(false);
   const [tooltipOpen, setTooltipOpen] = React.useState(false);
   const anchorRef = React.useRef(null);
-  const t = useSelector((state) => state.options.t);
+  const t = useTranslate();
   const dispatch = useDispatch();
   const userLanguage = useSelector((state) => state.options.userLanguage);
   const messages = useSelector((state) => state.notifications.messages);
