+++ docs/src/modules/redux/optionsReducer.js
@@ -1,42 +1,4 @@
 import { ACTION_TYPES, CODE_VARIANTS } from 'docs/src/modules/constants';
-import mapTranslations from 'docs/src/modules/utils/mapTranslations';
-
-const req = require.context('docs/translations', false, /translations.*\.json$/);
-const translations = mapTranslations(req, 'json');
-
-function getPath(obj, path) {
-  if (!path || typeof path !== 'string') {
-    return null;
-  }
-
-  return path.split('.').reduce((acc, item) => (acc && acc[item] ? acc[item] : null), obj);
-}
-
-const warnOnce = {};
-
-const getT = (userLanguage) => (key, options = {}) => {
-  const { ignoreWarning = false } = options;
-  const wordings = translations[userLanguage];
-
-  if (!wordings) {
-    console.error(`Missing language: ${userLanguage}.`);
-    return '…';
-  }
-
-  const translation = getPath(wordings, key);
-
-  if (!translation) {
-    const fullKey = `${userLanguage}:${key}`;
-    // No warnings in CI env
-    if (!ignoreWarning && !warnOnce[fullKey] && typeof window !== 'undefined') {
-      console.error(`Missing translation for ${fullKey}.`);
-      warnOnce[fullKey] = true;
-    }
-    return getPath(translations.en, key);
-  }
-
-  return translation;
-};
 
 const mapping = {
   [ACTION_TYPES.OPTIONS_CHANGE]: (state, action) => {
@@ -61,8 +23,6 @@ export default function optionsReducer(state = {}, action) {
   if (mapping[action.type]) {
     newState = mapping[action.type](state, action);
   }
-
-  newState.t = getT(newState.userLanguage);
 
   return newState;
 }
