+++ docs/src/pages/customization/density/DensityTool.js
@@ -10,7 +10,7 @@ import Switch from '@material-ui/core/Switch';
 import { DispatchContext } from 'docs/src/modules/components/ThemeContext';
 import IncreaseIcon from '@material-ui/icons/AddCircleOutline';
 import DecreaseIcon from '@material-ui/icons/RemoveCircleOutline';
-import { useSelector } from 'react-redux';
+import { useTranslate } from 'docs/src/modules/utils/i18n';
 
 const minSpacing = 0;
 const maxSpacing = 20;
@@ -46,7 +46,7 @@ export default function DensityTool() {
   const theme = useTheme();
   const spacingUnit = theme.spacing(1);
 
-  const t = useSelector((state) => state.options.t);
+  const t = useTranslate();
 
   return (
     <Grid container spacing={2}>
